add new product

+++ eclipse-workspace/HappyGraden/src/main/java/vn/com/unit/service/ProductImg2DService.java
@@ -2,5 +2,5 @@ package vn.com.unit.service;
 
 public interface ProductImg2DService {
 
-	public void saveImg2D(Long id, String url);
+	public void saveImg2D(int id, String url);
 }
